+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,12 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import open_mail_launcher
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  OpenMailLauncherPlugin.register(with: registry.registrar(forPlugin: "OpenMailLauncherPlugin"))
+}