+++ macos/open_mail_launcher/Package.swift
@@ -1,0 +1,28 @@
+// swift-tools-version: 5.9
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+    name: "open_mail_launcher",
+    platforms: [
+        .macOS("10.15")
+    ],
+    products: [
+        .library(name: "open-mail-launcher", targets: ["open_mail_launcher"])
+    ],
+    dependencies: [
+        .package(name: "FlutterFramework", path: "../FlutterFramework")
+    ],
+    targets: [
+        .target(
+            name: "open_mail_launcher",
+            dependencies: [
+                .product(name: "FlutterFramework", package: "FlutterFramework")
+            ],
+            resources: [
+                .process("PrivacyInfo.xcprivacy")
+            ]
+        )
+    ]
+)